+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [3.0.4] - 2024-08-20
 - Removed the get_inst function.
+- Added dict_out function to gen_class module.
+- Added Pretty Print to gen_libs.print_data function.
+
+### Added
+- gen_class.dict_out: Outputs the dictionary in a variety of formats and media.
 
 ### Changed
+- gen_libs.print_data:  Added Pretty Print cability to function.
 - gen_class.Daemon.stop, gen_class.Daemon2.stop, gen_class.Mail.read_stdin, gen_Class.Mail.send_mail, gen_Class.Mail.send_mailx, gen_libs.make_md5_hash, gen_libs.compress: Removed get_inst call.
 
 ### Removed